requirements.txt
openai-agents>=0.0.16
openai>=1.81.0
pydantic>=2.10
python-dotenv>=1.0.0 